Fix: "At least one of ratingCount or reviewCount is required."
Initially if you enable the plugin with Google Rich Snippet support and
you don't have any reviews yet the markup is not valid. The
AggregateRating object is created, but it has two missing required
fields:

- Field ratingValue may not be empty.
- At least one of ratingCount or reviewCount is required.

If we have no score yet, we do not create the AggregateRating object at
all.

Closes #64